feat: auth context
